% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_date.R
\name{convert_to_date}
\alias{convert_to_date}
\title{Convert a column to Date format}
\usage{
convert_to_date(data, column, format = "\%Y-\%m-\%d")
}
\arguments{
\item{data}{A data set containing a column to be converted.}

\item{column}{The name of the column to be converted.}

\item{format}{The date format (default is "%Y-%m-%d").}
}
\value{
A data set with the specified column converted to Date format.
}
\description{
This function converts a specified column in a dataset to Date format.
}
