% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_na.R
\name{fill_na}
\alias{fill_na}
\title{Fills in missing values}
\usage{
fill_na(data, value)
}
\arguments{
\item{data}{The data set with mising values}

\item{value}{The specifed value that is wanted to replace NA's}
}
\value{
returns the data set with filled in value
}
\description{
This function fills in mising values with a specified value
}
