% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imput_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Impute missing values}
\usage{
impute_missing(data, method = "mean")
}
\arguments{
\item{data}{A data set to be cleaned.}

\item{method}{The imputation method: "mean" or "median" (default is "mean").}
}
\value{
A data set with missing values imputed.
}
\description{
This function imputes missing values in numeric columns using the mean or median.
}
