% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicate rows}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{data set that has duplicate rows}
}
\value{
returns clean data set
}
\description{
Looks at data set and removes rows that are duplicates
}
