% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Removes na from data set}
\usage{
remove_na(data)
}
\arguments{
\item{data}{A data set that has NA values}
}
\value{
returns data set without NA vlaues
}
\description{
Takes in any data set and removes any missing values.
}
