% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized_names.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardizes column names}
\usage{
standardize_names(data)
}
\arguments{
\item{data}{desired dataset}
}
\value{
Data set with modifed names
}
\description{
Gets rid of any extra spaces, commas or underscores in column names.
}
