% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_strings.R
\name{trim_strings}
\alias{trim_strings}
\title{Trim whitespace from character columns}
\usage{
trim_strings(data)
}
\arguments{
\item{data}{A data set to be cleaned.}
}
\value{
A data set with whitespaces trimmed from character columns.
}
\description{
This function trims leading and trailing whitespaces from character columns.
}
